+++ test/standardized.jl
@@ -19,6 +19,7 @@ using RestrictedBoltzmannMachines: standardize, unstandardize, standardize!, uns
 using RestrictedBoltzmannMachines: standardize_hidden, standardize_visible
 using StatsBase: proportionmap
 using Statistics: mean
+using FillArrays: Ones
 using Test: @inferred, @test, @testset
 using Zygote: gradient
 
@@ -288,7 +289,7 @@ end
             (
                 binary_data,
                 potts_data,
-                nothing,
+                Ones{Bool}(4),
                 [0.0, 0.5],
                 [1.0, 0.5],
                 reshape([0.5, 0.5, 0.0], 3, 1),